refactor: create a hint widget
