+++ src/features/hints/index.ts
@@ -4,5 +4,3 @@ export {
 } from "./model/HintSettings";
 
 export { default as HintControls } from "./ui/HintControls";
-export { default as PronunciationHint } from "./ui/PronunciationHint";
-export { default as TranslationHint } from "./ui/TranslationHint";
