+++ src/widgets/hints/index.ts
@@ -1,0 +1,3 @@
+import Hints from "./ui/Hints";
+
+export default Hints;
