+++ src/widgets/hints/ui/TranslationHint.module.css
