+++ src/widgets/hints/ui/TranslationHint.ts
@@ -1,8 +1,6 @@
 import { Component } from "../../../shared/ui";
-
-import RoundState from "../../game/model/RoundState";
-import HintSettings from "../model/HintSettings";
-
+import { RoundState } from "../../../features/game";
+import { HintSettings } from "../../../features/hints";
 import { Observer, Publisher } from "../../../entities/state";
 
 import styles from "./TranslationHint.module.css";
